Transform axios instance to a hook and setup correct accept language header for all axios request
