+++ src/auth/hooks/useForgotPassword.ts
@@ -1,11 +1,15 @@
 import { useMutation } from '@tanstack/react-query';
-import axios from 'core/config/axios';
+import { AxiosInstance } from 'axios';
+import useAxios from 'core/hooks/useAxios';
+
+let axios: AxiosInstance;
 
 const forgotPassword = async ({ email }: { email: string }): Promise<void> => {
   await axios.post('/forgot-password', { email });
 };
 
 export function useForgotPassword() {
+  axios = useAxios();
   const { isLoading, mutateAsync } = useMutation(forgotPassword);
   return { isSending: isLoading, forgotPassword: mutateAsync };
 }
