+++ src/auth/hooks/useRegister.ts
@@ -1,5 +1,8 @@
 import { useMutation } from '@tanstack/react-query';
-import axios from 'core/config/axios';
+import { AxiosInstance } from 'axios';
+import useAxios from 'core/hooks/useAxios';
+
+let axios: AxiosInstance;
 
 const register = async (userData: {
   firstName: string;
@@ -13,6 +16,7 @@ const register = async (userData: {
 };
 
 export const useRegister = () => {
+  axios = useAxios();
   const { isLoading, mutateAsync } = useMutation(register);
   return { isRegistering: isLoading, register: mutateAsync };
 };
