+++ src/matrix/hooks/useCreateMatrix.ts
@@ -1,7 +1,10 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import axios from 'core/config/axios';
+import { AxiosInstance } from 'axios';
+import useAxios from 'core/hooks/useAxios';
 import { addOne } from 'core/utils/crudUtils';
 import Matrix from 'matrix/types/MatrixData';
+
+let axios: AxiosInstance;
 
 const createMatrix = async (matrix: Partial<Matrix>): Promise<Matrix> => {
   const { data } = await axios.post('/matrices', matrix);
@@ -9,6 +12,7 @@ const createMatrix = async (matrix: Partial<Matrix>): Promise<Matrix> => {
 };
 
 export const useCreateMatrix = () => {
+  axios = useAxios();
   const queryClient = useQueryClient();
 
   const { isLoading, mutateAsync } = useMutation(createMatrix, {
