+++ src/matrix/hooks/useDeleteMatrix.ts
@@ -1,7 +1,10 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import axios from 'core/config/axios';
+import { AxiosInstance } from 'axios';
+import useAxios from 'core/hooks/useAxios';
 import { removeOne } from 'core/utils/crudUtils';
 import Matrix from 'matrix/types/MatrixData';
+
+let axios: AxiosInstance;
 
 const deleteMatrix = async (id: number): Promise<Matrix> => {
   const { data } = await axios.delete(`/matrices/${id}`);
@@ -9,6 +12,7 @@ const deleteMatrix = async (id: number): Promise<Matrix> => {
 };
 
 export const useDeleteMatrix = () => {
+  axios = useAxios();
   const queryClient = useQueryClient();
 
   const { isLoading, mutateAsync } = useMutation(deleteMatrix, {
