+++ src/matrix/hooks/useGetMatrix.ts
@@ -1,6 +1,9 @@
 import { useQuery } from '@tanstack/react-query';
-import axios from 'core/config/axios';
+import { AxiosInstance } from 'axios';
+import useAxios from 'core/hooks/useAxios';
 import Matrix from 'matrix/types/MatrixData';
+
+let axios: AxiosInstance;
 
 const fetchMatrix = async (matrixId: number): Promise<Matrix | null> => {
   try {
@@ -12,6 +15,7 @@ const fetchMatrix = async (matrixId: number): Promise<Matrix | null> => {
 };
 
 export const useGetMatrix = (matrixId: number) => {
+  axios = useAxios();
   const { data } = useQuery(['matrix', matrixId], () => fetchMatrix(matrixId));
   return { data: data ?? undefined };
 };
