+++ src/session/hooks/useJoinSession.ts
@@ -1,7 +1,10 @@
 import { useMutation } from '@tanstack/react-query';
-import axios from 'core/config/axios';
+import { AxiosInstance } from 'axios';
+import useAxios from 'core/hooks/useAxios';
 import MatrixData from 'matrix/types/MatrixData';
 import SessionData from 'session/types/SessionData';
+
+let axios: AxiosInstance;
 
 const joinSession = async (
   hashId: string
@@ -11,6 +14,7 @@ const joinSession = async (
 };
 
 export const useJoinSession = () => {
+  axios = useAxios();
   const { isLoading, mutateAsync } = useMutation(joinSession);
   return { isJoining: isLoading, joinSession: mutateAsync };
 };
