+++ src/user/hooks/useUpdateName.ts
@@ -1,6 +1,9 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import axios from 'core/config/axios';
+import { AxiosInstance } from 'axios';
+import useAxios from 'core/hooks/useAxios';
 import UserData from 'user/types/userData';
+
+let axios: AxiosInstance;
 
 const updateName = async ({
   firstName,
@@ -9,11 +12,15 @@ const updateName = async ({
   firstName: string;
   lastName: string;
 }): Promise<UserData> => {
-  const { data } = await axios.patch('/users/name', { firstName, lastName });
+  const { data } = await axios.patch('/users/name', {
+    firstName,
+    lastName,
+  });
   return data.data.user;
 };
 
 export function useUpdateName(accessToken: string) {
+  axios = useAxios();
   const queryClient = useQueryClient();
 
   const { isLoading, mutateAsync } = useMutation(updateName, {
