+++ src/user/hooks/useUpdatePassword.ts
@@ -1,5 +1,8 @@
 import { useMutation } from '@tanstack/react-query';
-import axios from 'core/config/axios';
+import { AxiosInstance } from 'axios';
+import useAxios from 'core/hooks/useAxios';
+
+let axios: AxiosInstance;
 
 const updatePassword = async ({
   password,
@@ -18,6 +21,7 @@ const updatePassword = async ({
 };
 
 export function useUpdatePassword() {
+  axios = useAxios();
   const { isLoading, mutateAsync } = useMutation(updatePassword);
   return { isUpdating: isLoading, updatePassword: mutateAsync };
 }
